+++ advent/src/key_map.cxx
@@ -67,7 +67,7 @@ KeyMap::get_binding (int key)
 bool 
 KeyMap::is_a (SCM smob)
 {
-  return (int(SCM_CAR (smob)) == get_smob_tag ());
+  return (long(SCM_CAR (smob)) == get_smob_tag ());
 }
 
 void
